[impls] Move trait impls from crate root
